Replace RouterRenderer with higher order component
